pyproject.toml
[tool.poetry]
name = "proyecto1-bases-2"
version = "0.1.0"
description = ""
authors = ["JersonPQ <prendes.2507@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.3"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
pymongo = "^4.6.3"
redis = "^5.0.3"
pyjwt = "^2.8.0"
pytz = "^2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
